cleanup: Refactor ceph-related operations
The `microceph.ceph` operations are spread out in various modules.
Centralize them into `ceph.py`.
