+++ src/ceph.py
@@ -27,12 +27,21 @@ https://github.com/juju/charm-helpers/blob/master/charmhelpers/contrib/storage/l
 * only moved functions that are required by the charm
 """
 
+import collections
+import enum
+import functools
 import json
 import logging
 import math
 import os
 import socket
+import subprocess
 from subprocess import CalledProcessError, check_call, check_output
+from typing import Dict, List, Tuple, TypeAlias
+
+from tenacity import retry, stop_after_attempt, wait_fixed
+
+import utils
 
 CRITICAL = "CRITICAL"
 ERROR = "ERROR"
@@ -57,6 +66,8 @@ AUTOSCALER_DEFAULT_PGS = 32
 LEADER = "leader"
 PEON = "peon"
 QUORUM = [LEADER, PEON]
+
+VAR_LIB_CEPH = "/var/snap/microceph/common/data"
 
 logger = logging.getLogger(__name__)
 
@@ -124,6 +135,117 @@ def validator(value, valid_type, valid_range=None):
             assert (
                 value <= valid_range[1]
             ), "{} is greater than maximum allowed value of {}".format(value, valid_range[1])
+
+
+# Below functions are picked from
+# https://opendev.org/openstack/charms.ceph/src/branch/master/charms_ceph/utils.py
+
+Capabilities: TypeAlias = Dict[str, List[str]]
+
+_default_caps: Capabilities = collections.OrderedDict(
+    [
+        ("mon", ["allow r", 'allow command "osd blacklist"', 'allow command "osd blocklist"']),
+        ("osd", ["allow rwx"]),
+    ]
+)
+
+
+def parse_key(raw_key):
+    """Parse the key."""
+    # get-or-create appears to have different output depending
+    # on whether its 'get' or 'create'
+    # 'create' just returns the key, 'get' is more verbose and
+    # needs parsing
+    key = None
+    if len(raw_key.splitlines()) == 1:
+        key = raw_key
+    else:
+        for element in raw_key.splitlines():
+            if "key" in element:
+                return element.split(" = ")[1].strip()  # IGNORE:E1103
+    return key
+
+
+@functools.lru_cache()
+def ceph_auth_get(key_name):
+    """Get ceph auth key."""
+    try:
+        # Does the key already exist?
+        output = str(
+            check_output(
+                [
+                    "microceph.ceph",
+                    "--name",
+                    "mon.",
+                    "--keyring",
+                    f"{VAR_LIB_CEPH}/mon/ceph-{socket.gethostname()}/keyring",
+                    "auth",
+                    "get",
+                    key_name,
+                ]
+            ).decode("UTF-8")
+        ).strip()
+        return parse_key(output)
+    except CalledProcessError:
+        # Couldn't get the key
+        pass
+
+
+def get_named_key(name, caps=None, pool_list=None):
+    """Retrieve a specific named cephx key.
+
+    :param name: String Name of key to get.
+    :param pool_list: The list of pools to give access to
+    :param caps: dict of cephx capabilities
+    :returns: Returns a cephx key
+    """
+    key = ceph_auth_get(name)
+    if key:
+        return key
+
+    log("Creating new key for {}".format(name), level=DEBUG)
+    caps = caps or _default_caps
+    cmd = [
+        "microceph.ceph",
+        "--name",
+        "mon.",
+        "--keyring",
+        f"{VAR_LIB_CEPH}/mon/ceph-{socket.gethostname()}/keyring",
+        "auth",
+        "get-or-create",
+        name,
+    ]
+    # Add capabilities
+    for subsystem, subcaps in caps.items():
+        if subsystem == "osd":
+            if pool_list:
+                # This will output a string similar to:
+                # "pool=rgw pool=rbd pool=something"
+                pools = " ".join(["pool={0}".format(i) for i in pool_list])
+                subcaps[0] = subcaps[0] + " " + pools
+        cmd.extend([subsystem, "; ".join(subcaps)])
+    ceph_auth_get.cache_clear()
+
+    log("Calling check_output: {}".format(cmd), level=DEBUG)
+    return parse_key(str(check_output(cmd).decode("UTF-8")).strip())  # IGNORE:E1103
+
+
+def is_leader():
+    """Check if this node is ceph mon leader."""
+    hostname = socket.gethostname()
+    cmd = ["microceph.ceph", "tell", f"mon.{hostname}", "mon_status", "--format", "json"]
+    try:
+        result = json.loads(str(check_output(cmd).decode("UTF-8")))
+    except CalledProcessError:
+        return False
+    except ValueError:
+        # Non JSON response from mon_status
+        return False
+
+    if result["state"] == LEADER:
+        return True
+    else:
+        return False
 
 
 def monitor_key_get(service, key):
@@ -218,6 +340,88 @@ def update_pool(client, pool, settings):
         check_call(cmd + extend_cmd)
 
 
+def delete_pool(service, request):
+    """Delete a RADOS pool from ceph."""
+    cmd = [
+        "microceph.ceph",
+        "--id",
+        service,
+        "osd",
+        "pool",
+        "delete",
+        request.get("name"),
+        "--yes-i-really-really-mean-it",
+    ]
+    check_call(cmd)
+
+
+def rename_pool(service, request):
+    """Rename a Ceph pool from old_name to new_name.
+
+    :param service: The Ceph user name to run the command under.
+    :type service: str
+    :param request: The request with the old and new names for the pool.
+    :type request: dict
+    """
+    cmd = [
+        "microceph.ceph",
+        "--id",
+        service,
+        "osd",
+        "pool",
+        "rename",
+        request.get("name"),
+        request.get("new-name"),
+    ]
+    check_call(cmd)
+
+
+def snapshot_pool(service, request):
+    """Snapshots a RADOS pool in Ceph.
+
+    :param service: The Ceph user name to run the command under.
+    :type service: str
+    :param request: The request with the pool and snapshot names.
+    :type snapshot_name: dict
+    :raises: CalledProcessError
+    """
+    cmd = [
+        "microceph.ceph",
+        "--id",
+        service,
+        "osd",
+        "pool",
+        "mksnap",
+        request.get("name"),
+        request.get("snapshot-name"),
+    ]
+    check_call(cmd)
+
+
+def remove_pool_snapshot(service, request):
+    """Remove a snapshot from a RADOS pool in Ceph.
+
+    :param service: The Ceph user name to run the command under.
+    :type service: str
+    :param pool_name: Name of pool to remove snapshot from.
+    :type pool_name: str
+    :param snapshot_name: Name of snapshot to remove.
+    :type snapshot_name: str
+    :raises: CalledProcessError
+    """
+    cmd = [
+        "microceph.ceph",
+        "--id",
+        service,
+        "osd",
+        "pool",
+        "rmsnap",
+        request.get("name"),
+        request.get("snapshot-name"),
+    ]
+    check_call(cmd)
+
+
 def set_app_name_for_pool(client, pool, name):
     """Calls `osd pool application enable` for the specified pool name.
 
@@ -248,6 +452,90 @@ def enabled_manager_modules():
         return []
     modules = json.loads(modules)
     return modules["enabled_modules"]
+
+
+@retry(wait=wait_fixed(5), stop=stop_after_attempt(10))
+def list_mgr_modules() -> dict:
+    """Returns a python dict of mgr modules.
+
+    available keys:
+       1. disabled_modules
+       2. always_on_modules
+       3. enabled_modules
+    """
+    cmd = ["microceph.ceph", "mgr", "module", "ls", "--format", "json"]
+    return json.loads(utils.run_cmd(cmd=cmd))
+
+
+def enable_mgr_module(module: str):
+    """Enable requested ceph mgr module."""
+    disabled_modules = [
+        module_info["name"] for module_info in list_mgr_modules()["disabled_modules"]
+    ]
+    if module not in disabled_modules:
+        logger.info("nothing to do, %s module is not disabled", module)
+        return
+
+    cmd = ["microceph.ceph", "mgr", "module", "enable", module]
+    utils.run_cmd(cmd=cmd)
+
+
+def disable_mgr_module(module: str):
+    """Disable requested ceph mgr module."""
+    enabled_modules = list_mgr_modules()["enabled_modules"]
+    if module not in enabled_modules:
+        logger.info("nothing to do, %s module is not enabled or is always on", module)
+        return
+
+    cmd = ["microceph.ceph", "mgr", "module", "disable", module]
+    utils.run_cmd(cmd=cmd)
+
+
+def enable_ceph_monitoring():
+    """Enable Monitoring for ceph cluster."""
+    enable_mgr_module("prometheus")
+
+
+def disable_ceph_monitoring():
+    """Disable Monitoring for ceph cluster."""
+    disable_mgr_module("prometheus")
+
+
+class CephHealth(enum.Enum):
+    """Enumerate ceph health status."""
+
+    Ok = "HEALTH_OK"
+    Warn = "HEALTH_WARN"
+    Err = "HEALTH_ERR"
+    Unknown = "HEALTH_UNKNOWN"
+
+    @classmethod
+    def from_string(cls, health_str: str):
+        """Construct a CephHealth object from a string."""
+        for health in cls:
+            if health.value == health_str:
+                return health
+        return cls.Unknown
+
+    def __str__(self):
+        """Return the string representation of the health."""
+        return self.value
+
+
+class CephStatus(object):
+    """Class to handle ceph health checks."""
+
+    def ceph_health(self) -> Tuple[CephHealth, str]:
+        """Return the health of the monitor."""
+        cmd = ["sudo", "microceph.ceph", "health", "detail", "--format=json"]
+        try:
+            output = utils.run_cmd(cmd)
+        except subprocess.CalledProcessError:
+            # ceph health detail command failed, possibly mon wasn't reachable
+            # as it's restarting. Return unknown health for this case.
+            return CephHealth.Unknown, "fault running ceph health detail command"
+        res = json.loads(output.strip())
+        return CephHealth.from_string(res["status"]), res["checks"]
 
 
 def enable_pg_autoscale(service, pool_name):
@@ -320,6 +608,32 @@ def get_osds(service, device_class=None):
             ["microceph.ceph", "--id", service, "osd", "ls", "--format=json"]
         ).decode("utf-8")
     return json.loads(out)
+
+
+def get_osd_weight(osd_id):
+    """Returns the weight of the specified OSD.
+
+    :returns: Float
+    :raises: ValueError if the monmap fails to parse.
+    :raises: CalledProcessError if our Ceph command fails.
+    """
+    try:
+        tree = check_output(["microceph.ceph", "osd", "tree", "--format=json"])
+        tree = tree.decode("UTF-8")
+        try:
+            json_tree = json.loads(tree)
+            # Make sure children are present in the JSON
+            if not json_tree["nodes"]:
+                return None
+            for device in json_tree["nodes"]:
+                if device["type"] == "osd" and device["name"] == osd_id:
+                    return device["crush_weight"]
+        except ValueError as v:
+            log("Unable to parse ceph tree json: {}. Error: {}".format(tree, v))
+            raise
+    except CalledProcessError as e:
+        log("ceph osd tree command failed with message: {}".format(e))
+        raise
 
 
 def get_erasure_profile(service, name):
